refactor: split middlewares implementation
