+++ server/middlewares/index.js
@@ -1,0 +1,9 @@
+const reqIdSetter = require('./req-id-setter');
+const reqLogger = require('./logger');
+const errorHandlers = require('./error');
+
+module.exports = {
+  reqIdSetter,
+  reqLogger,
+  errorHandlers,
+};
